feat(tooling): add releases
