+++ api/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "dlv": "^1.1.3",
     "merge-options": "^3.0.3",
-    "nanoid": "^3.1.16"
+    "nanoid": "^3.1.16",
+    "p-map": "^4.0.0"
   }
 }
