+++ api/webpack.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  target: 'webworker',
+  // devtool: 'cheap-module-source-map',
+  entry: './index.js',
+  mode: 'production',
+}
