+++ api/wrangler.toml
@@ -4,3 +4,4 @@ account_id = "fffa4b4363a7e5250af8357087263b3a"
 workers_dev = true
 route = ""
 zone_id = ""
+webpack_config = "webpack.config.js"
