+++ content/releases/_index.md
@@ -1,0 +1,8 @@
+---
+title: Releases
+weight: 9
+---
+
+# Releases
+
+{{<changelog>}}
